[APPS-0014] connect the dots, make ui look better so that it will not hurt your eyes, mapping data

+++ weather-forecast/src/main/java/vn/nab/innovation/center/assignment/android/tungphan/weather/forecast/data/datasource/remote/dto/MainDTO.kt
@@ -30,5 +30,5 @@ data class MainDTO(
     var humidity: Int? = null,
     @SerializedName("temp_kf")
     @Expose
-    var tempKf: Int? = null
+    var tempKf: Double? = null
 )
