+++ weather-forecast/src/main/java/vn/nab/innovation/center/assignment/android/tungphan/weather/forecast/domain/entity/Main.kt
@@ -9,5 +9,5 @@ data class Main(
     var seaLevel: Int? = null,
     var grndLevel: Int? = null,
     var humidity: Int? = null,
-    var tempKf: Int? = null
+    var tempKf: Double? = null
 )
